refactor base hook; remove errant dep
